상직 - 20170825 SQLite 완료, 설문 후 설문이력에 추가

+++ app/src/main/java/com/landvibe/goodbeing/goodbeing_android/History/Adapter/HistorySurveyListAdapter.java
@@ -55,7 +55,7 @@ public class HistorySurveyListAdapter extends RecyclerView.Adapter<HistorySurvey
         holder.survey_item_score.setText(item.getSurvey_score() + "점");
         holder.survey_item_date.setText(item.getSurvey_date());
 
-        int score = Integer.parseInt(item.getSurvey_score());
+        float score = Float.parseFloat(item.getSurvey_score());
         String result = new String();
 
         if(score > 70)
